Fixed casting warnings.

+++ mz_strm.c
@@ -284,7 +284,7 @@ int32_t mz_stream_find(void *stream, const void *find, int32_t find_size, int64_
 
     while (read_pos < max_seek)
     {
-        if (read_size > (int32_t)(max_seek - read_pos - buf_pos) && (max_seek - read_pos - buf_pos) < sizeof(buf))
+        if (read_size > (int32_t)(max_seek - read_pos - buf_pos) && (max_seek - read_pos - buf_pos) < (int64_t)sizeof(buf))
             read_size = (int32_t)(max_seek - read_pos - buf_pos);
 
         read = mz_stream_read(stream, buf + buf_pos, read_size);
@@ -346,7 +346,7 @@ int32_t mz_stream_find_reverse(void *stream, const void *find, int32_t find_size
 
     while (read_pos < max_seek)
     {
-        if (read_size > (int32_t)(max_seek - read_pos) && (max_seek - read_pos) < sizeof(buf))
+        if (read_size > (int32_t)(max_seek - read_pos) && (max_seek - read_pos) < (int64_t)sizeof(buf))
             read_size = (int32_t)(max_seek - read_pos);
 
         if (mz_stream_seek(stream, start_pos - (read_pos + read_size), MZ_SEEK_SET) != MZ_OK)
